Fix the Okhttp Otel parent span feature when not using RUM

+++ integrations/dd-sdk-android-okhttp-otel/src/test/kotlin/com/datadog/android/okhttp/otel/OkHttpExtTest.kt
@@ -7,8 +7,8 @@
 package com.datadog.android.okhttp.otel
 
 import com.datadog.android.okhttp.TraceContext
+import com.datadog.legacy.trace.api.sampling.PrioritySampling
 import com.datadog.tools.unit.forge.BaseConfigurator
-import com.datadog.trace.api.sampling.PrioritySampling
 import fr.xgouchet.elmyr.annotation.BoolForgery
 import fr.xgouchet.elmyr.annotation.StringForgery
 import fr.xgouchet.elmyr.junit5.ForgeConfiguration
@@ -61,7 +61,7 @@ internal class OkHttpExtTest {
             on { isSampled }.thenReturn(fakeIsSampled)
         }
         expectedPrioritySampling =
-            if (fakeIsSampled) PrioritySampling.USER_KEEP.toInt() else PrioritySampling.UNSET.toInt()
+            if (fakeIsSampled) PrioritySampling.USER_KEEP else PrioritySampling.UNSET
         whenever(mockSpan.spanContext).thenReturn(spanContext)
     }
